requirements.txt
fastapi==0.104.1
uvicorn==0.24.0.post1
pydantic==2.5.2
pandas==2.1.4
openai==1.3.7  # Still included for optional use
google-generativeai==0.7.2  # Added for Gemini (check latest version)
pinecone-client==2.2.4
python-dotenv==1.0.0
python-multipart==0.0.5